saferide-backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
geopy==2.4.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
slowapi==0.1.9
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
alembic==1.12.1
email-validator==2.2.0 